updated readMe file to to show travis status

+++ main/business/businessViews.py
@@ -1,10 +1,11 @@
 from flask import Blueprint, Flask, request, json, jsonify, make_response, url_for, abort
 import jwt
 import datetime
-from user.userViews import token_required, loggedInUser
+from datetime import datetime
+from ..user.userViews import token_required, logged_in_user
 from werkzeug.security import generate_password_hash, check_password_hash
 from flasgger import swag_from
-from appModels import Business, db
+from ..appModels import Business, db
 from flask_paginate import Pagination, get_page_args, get_page_parameter
 # from ..run import businessBlueprint
 
@@ -14,14 +15,38 @@ businessBlueprint = Blueprint('business', __name__)
 @businessBlueprint.route('/api/businesses', methods=['POST'])
 @swag_from('createBusiness.yml')
 # @token_required
-def createBusiness():
-    global loggedInUser
+def create_business():
+    global logged_in_user
     jsn = request.data
     data = json.loads(jsn)
 
+    if len(data.keys()) == 0:
+        return jsonify({'message':'fill in all the fields that is name, location, category and description'}), 400 #bad request
+
+    if len(data.keys()) != 4:
+        return jsonify({'message':'cannot create business because of missing fields'}), 400 #bad request
+    
+    if 'name' not in data.keys():
+        return jsonify({'message':'business name is missing'}), 400 #bad request
+
+    if 'location' not in data.keys():
+        return jsonify({'message':'location is missing'}), 400 #bad request
+
+    if 'category' not in data.keys():
+        return jsonify({'message':'category is missing'}), 400 #bad request
+
+    if 'description' not in data.keys():
+        return jsonify({'message':'description is missing'}), 400 #bad request
+
+    if len(data['name']) < 5:
+        return jsonify({'message':'name of business is too short should between five and ten characters'}), 400 #bad request
+
+    if len(data['name']) > 10:
+        return jsonify({'message':'name of business is too long should between five and ten characters'}), 400 #bad request
+
     #lets pick the data from json passed
     bizname = data['name']
-    userid = loggedInUser['id']
+    userid = logged_in_user['id']
     location =data['location']
     category = data['category']
     description = data['description']
@@ -45,7 +70,7 @@ def createBusiness():
 @businessBlueprint.route('/api/businesses/<int:id>', methods=['GET'])
 @swag_from('retrieveBusiness.yml')
 # @token_required
-def getOneBusiness(id):
+def get_one_business(id):
     biz = Business.query.get(id)
     if biz:
         return jsonify({'message':biz.returnJson()}), 200
@@ -58,82 +83,68 @@ def getOneBusiness(id):
 @businessBlueprint.route('/api/businesses', methods=['GET'])
 @swag_from('retrieveAllBusinesses.yml')
 # @token_required
-def getAllBusinesses():
-    businesses = Business.query.filter_by(id=2)
-    # print(Business.query.filter_by(id=1))
-    # print(businesses.count())
+def get_all_businesses():
+    businesses = Business.query.all()    
     print(type(businesses))
     if not businesses:
-        return jsonify({'message':'No businesses exist, please register one'}), 404
+        return jsonify({'message':'no businesses exist, please register one'}), 404
     else:
-        
-        page_num = request.args.get('page', 1, type=int)
-        limit = request.args.get('limit', 1, type=int)
-        item = Business.query.paginate(per_page=1, page=page_num, error_out=False)        
-        results = item.items
-        
-        prev =None
-        if item.has_prev:
-            # prev = url_for('/api/businesses?page={}&limit={}'.format(page_num, limit), page=page_num-1, _external=True)
-            prev = url_for('127.0.0.1:5000/api/businesses', page=page_num-1, _external=True)
-        next=None
-        if item.has_next:
-            # next = url_for('/api/businesses?page={}&limit={}'.format(page_num, limit), page=page_num+1, _external=True)
-            next = url_for('127.0.0.1:5000/api/businesses', page=page_num-1, _external=True)
-        return jsonify({'businesses':[business.returnJson() for business in results],'previous':prev, 'next':next}), 200        
+        return jsonify({"businesses":[business.returnJson() for business in businesses]}), 200
         
 
 @businessBlueprint.route('/api/businesses/<int:id>', methods=['PUT'])
 @swag_from('updateBusiness.yml')
 # @token_required
-def updatebusiness(id):
-    global loggedInUser
-    biz = Business.query.get(id)
-    print(biz.id)
+def update_business(id):
+    global logged_in_user
+    jsn = request.data
+    data = json.loads(jsn)
+    biz = Business.query.get(int(id))    
+    
+    if not logged_in_user:
+        return jsonify({"message":"please login"})
 
-    if len(loggedInUser.keys()) != 3:
-        return jsonify({'message':'You are not logged in.'})
-    else:
-        userid = loggedInUser['id']
+    if 'id' not in logged_in_user.keys():
+        return jsonify({"message":"missing userid, cannot update business"})    
 
-    if biz.query.count() > 0:
-        print(loggedInUser)
-        
-        jsn = request.data
-        data = json.loads(jsn)
+    userid = logged_in_user['id']
+    
+    if biz.query.count() > 0 and userid == biz.userid:
+        # print(logged_in_user)
 
         if 'name' in data.keys():
             bizname = data['name']
         else:
-            bizname = None
+            bizname = ''
 
         if 'location' in data.keys():
             location =data['location']
         else:
-            location = None
+            location = ''
 
         if 'category' in data.keys():
             category = data['category']
         else:
-            category = None
+            category = ''
 
         if 'description' in data.keys():
             description = data['description']
         else:
-            description = None        
-
-        if biz.userid == str(userid):
+            description = ''        
             
-            if bizname:
-                biz.bizname = bizname
-            if location:
-                biz.location = location
-            if category:
-                biz.category = category
-            if description:
-                biz.description = description            
-            print([bizname, location, category, description])
-            db.session.add(biz)
+        if bizname:
+            biz.bizname = bizname
+        if location:
+            biz.location = location
+        if category:
+            biz.category = category
+        if description:
+            biz.description = description   
+
+        biz.date_modified = datetime.now()
+        print([bizname, location, category, description])
+        db.session.add(biz)
+        db.session.commit()
         return jsonify({'message':'business '+ str(id) +' has been updated successfully'}), 200
         
     else:
@@ -143,14 +154,22 @@ def updatebusiness(id):
 @businessBlueprint.route('/api/businesses/<int:id>', methods=['DELETE'])
 @swag_from('deleteBusiness.yml')
 # @token_required
-def deletebusiness(id):
-    global loggedInUser
-    userid = loggedInUser['id']
-    biz = Business.query.get(id)
-    print(userid == biz.id)
+def delete_business(id):
+    global logged_in_user    
+    biz = Business.query.get(int(id))
+
+    if not logged_in_user:
+        return jsonify({"message": "please login"}), 400
+
+    if 'id' not in logged_in_user.keys():
+        return jsonify({"message": "missing userid, cannot delete buiness"})
+
+    userid = logged_in_user['id']
+    
     if biz:
         if str(userid) == str(biz.id):
             db.session.delete(biz)
+            db.session.commit()
             return jsonify({'message':'business was deleted successfully'})
         else:
             return jsonify({'message':'business was not deleted because you are not the owner'})
@@ -161,32 +180,44 @@ def deletebusiness(id):
 @businessBlueprint.route('/api/businesses/search', methods=['GET'])
 @swag_from('searchBusiness.yml')
 # @token_required
-def searchBusiness():
+def search_business():
     name = request.args.get('q')
-    location = request.args.get('location')
-    category = request.args.get('category')
-    #/api/businesses/search?q=<business name>&filtery_type=<location or category>&filter_value=<value>
-    filter_type = request.args.get('filter_type')
-    filter_value = request.args.get('filter_value')
+    filter_type = str(request.args.get('filter_type'))
+    filter_value = str(request.args.get('filter_value'))
+
+    results = Business.query.filter_by(bizname=Business.bizname.ilike('%' + name + '%'))
 
     if filter_type == 'location':
-        result = Business.query.filter_by(bizname=name, location=filter_value)
+        results = Business.query.filter_by(bizname=name, location=filter_value).filter(Business.bizname.like("%"+ name +"%"))
     
     elif filter_type == 'category':
-        result = Business.query.filter_by(bizname=name, category=filter_value)
+        results = Business.query.filter_by(bizname=name, category=filter_value).filter(Business.bizname.like("%"+ name +"%"))
     else:
-        return jsonify({'message':'unknown filter type passed in query url'}), 400
+        return jsonify({'message':'no or unknown filter type passed in query url'}), 400
 
-    if result.count() > 0:
-        print(result)
-        return jsonify({'businesses':[res.returnJson() for res in result]}), 200
+    print(results)
+    if results.count() > 0:
+        print(results)
+        return jsonify({'businesses':[res.returnJson() for res in results]}), 200
     else:
-        return jsonify({'message':'No businesses match your search'}), 404
+        return jsonify({'message':'no businesses match your search'}), 404
 
 
 @businessBlueprint.route('/api/businesses/filter', methods=['GET'])
 @swag_from('filterBusiness.yml')
 # @token_required
-def filterBusiness():
-    pass
+def filter_business():
+    filter_type = str(request.args.get('filter_type'))
+    filter_value = str(request.args.get('filter_value'))
 
+    if filter_type == 'location':
+        results = Business.query.filter_by(location=filter_value)
+    elif filter_type == 'category':
+        results = Business.query.filter_by(category=filter_value)
+    else:
+        return jsonify({'message':'invalid or unknown filter type passed in query url'}), 400
+
+    if not results:
+        return jsonify({"message":"no businesses registered with that filter"}), 404
+
+    return jsonify({"businesses":[business.returnJson() for business in results]}), 200
